refactor: move data format utils to separate module
